requirements.txt
pip~=23.2.1
wheel~=0.41.2
setuptools~=68.2.0
Django~=4.2.7
asgiref~=3.7.2
sqlparse~=0.4.4
djangorestframework==3.14.0
drf-yasg==1.21.7